+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/&bloom.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>&Bloom - backoffice</title>
+    <title>&Bloom｜植物販賣所</title>
   </head>
   <body>
     <div id="root"></div>
